added code refractor for app.rb
